moved login to component since it is not a page
